feat(v11): add slot to layers
