+++ android/src/main/java/com/rnmapbox/rnmbx/components/styles/layers/codeparts/LayerManagerCommonProps.codepart-kt.ejs
@@ -1,0 +1,59 @@
+    @ReactProp(name = "id")
+    override fun setId(layer: <%= layerType %>, id: Dynamic) {
+        layer.iD = id.asString()
+    }
+
+    @ReactProp(name = "existing")
+    override fun setExisting(layer: <%= layerType %>, existing: Dynamic) {
+        layer.setExisting(existing.asBoolean())
+    }
+
+    @ReactProp(name = "sourceID")
+    override fun setSourceID(layer: <%= layerType %>, sourceID: Dynamic) {
+        layer.setSourceID(sourceID.asString())
+    }
+
+    @ReactProp(name = "aboveLayerID")
+    override fun setAboveLayerID(layer: <%= layerType %>, aboveLayerID: Dynamic) {
+        layer.setAboveLayerID(aboveLayerID.asString())
+    }
+
+    @ReactProp(name = "belowLayerID")
+    override fun setBelowLayerID(layer: <%= layerType %>, belowLayerID: Dynamic) {
+        layer.setBelowLayerID(belowLayerID.asString())
+    }
+
+    @ReactProp(name = "layerIndex")
+    override fun setLayerIndex(layer: <%= layerType %>, layerIndex: Dynamic) {
+        layer.setLayerIndex(layerIndex.asInt())
+    }
+
+    @ReactProp(name = "minZoomLevel")
+    override fun setMinZoomLevel(layer: <%= layerType %>, minZoomLevel: Dynamic) {
+        layer.setMinZoomLevel(minZoomLevel.asDouble())
+    }
+
+    @ReactProp(name = "maxZoomLevel")
+    override fun setMaxZoomLevel(layer: <%= layerType %>, maxZoomLevel: Dynamic) {
+        layer.setMaxZoomLevel(maxZoomLevel.asDouble())
+    }
+
+    @ReactProp(name = "reactStyle")
+    override fun setReactStyle(layer: <%= layerType %>, style: Dynamic) {
+        layer.setReactStyle(style.asMap())
+    }
+
+    @ReactProp(name = "sourceLayerID")
+    override fun setSourceLayerID(layer: <%= layerType %>, sourceLayerID: Dynamic) {
+        layer.setSourceLayerID(sourceLayerID.asString())
+    }
+
+    @ReactProp(name = "filter")
+    override fun setFilter(layer: <%= layerType %>, filterList: Dynamic) {
+        layer.setFilter(filterList.asArray())
+    }
+
+    @ReactProp(name = "slot")
+    override fun setSlot(layer: <%= layerType %>, slot: Dynamic) {
+        layer.setSlot(slot.asString())
+    }
