+++ android/src/main/mapbox-v11-compat/v10/com/rnmapbox/rnmbx/v11compat/Layer.kt
@@ -1,0 +1,7 @@
+package com.rnmapbox.rnmbx.v11compat.layer;
+
+import com.mapbox.maps.extension.style.layers.Layer
+
+fun Layer.slot(slot: String) {
+    // V11 only
+}
