+++ android/src/main/mapbox-v11-compat/v11/com/rnmapbox/rnmbx/v11compat/Layer.kt
@@ -1,0 +1,3 @@
+package com.rnmapbox.rnmbx.v11compat.layer;
+
+import com.mapbox.maps.extension.style.layers.Layer
