+++ docs/CircleLayer.md
@@ -42,7 +42,7 @@ The source from which to obtain the data to style.
 If the source has not yet been added to the current style, the behavior is undefined.
 Inferred from parent source only if the layer is a direct child to it.
 
-  _defaults to:_ `MapboxGL.StyleSource.DefaultSourceID`
+  _defaults to:_ `Mapbox.StyleSource.DefaultSourceID`
 
   
 ### sourceLayerID
@@ -50,8 +50,7 @@ Inferred from parent source only if the layer is a direct child to it.
 ```tsx
 string
 ```
-Identifier of the layer within the source identified by the sourceID property
-from which the receiver obtains the data to style.
+Identifier of the layer within the source identified by the sourceID property from which the receiver obtains the data to style.
 
 
   
@@ -106,6 +105,17 @@ The minimum zoom level at which the layer gets parsed and appears.
 number
 ```
 The maximum zoom level at which the layer gets parsed and appears.
+
+
+  
+### slot
+
+```tsx
+'bottom' | 'middle' | 'top'
+```
+The slot this layer is assigned to. If specified, and a slot with that name exists, it will be placed at that position in the layer order.
+
+v11 only
