+++ docs/RasterLayer.md
@@ -20,7 +20,7 @@ RasterLayer
 string
 ```
 _required_
-A string that uniquely identifies the layer in the style to which it is added.
+A string that uniquely identifies the source in the style to which it is added.
 
 
   
@@ -42,7 +42,7 @@ The source from which to obtain the data to style.
 If the source has not yet been added to the current style, the behavior is undefined.
 Inferred from parent source only if the layer is a direct child to it.
 
-  _defaults to:_ `MapboxGL.StyleSource.DefaultSourceID`
+  _defaults to:_ `Mapbox.StyleSource.DefaultSourceID`
 
   
 ### sourceLayerID
@@ -105,6 +105,17 @@ The minimum zoom level at which the layer gets parsed and appears.
 number
 ```
 The maximum zoom level at which the layer gets parsed and appears.
+
+
+  
+### slot
+
+```tsx
+'bottom' | 'middle' | 'top'
+```
+The slot this layer is assigned to. If specified, and a slot with that name exists, it will be placed at that position in the layer order.
+
+v11 only
