+++ docs/SymbolLayer.md
@@ -20,7 +20,7 @@ SymbolLayer is a style layer that renders icon and text labels at points or alon
 string
 ```
 _required_
-A string that uniquely identifies the layer in the style to which it is added.
+A string that uniquely identifies the source in the style to which it is added.
 
 
   
@@ -42,7 +42,7 @@ The source from which to obtain the data to style.
 If the source has not yet been added to the current style, the behavior is undefined.
 Inferred from parent source only if the layer is a direct child to it.
 
-  _defaults to:_ `MapboxGL.StyleSource.DefaultSourceID`
+  _defaults to:_ `Mapbox.StyleSource.DefaultSourceID`
 
   
 ### sourceLayerID
@@ -84,7 +84,7 @@ Inserts a layer at a specified index
 ### filter
 
 ```tsx
-Expression
+FilterExpression
 ```
 Filter only the features in the source layer that satisfy a condition that you define
 
@@ -108,12 +108,24 @@ The maximum zoom level at which the layer gets parsed and appears.
 
 
   
+### slot
+
+```tsx
+'bottom' | 'middle' | 'top'
+```
+The slot this layer is assigned to. If specified, and a slot with that name exists, it will be placed at that position in the layer order.
+
+v11 only
+
+
+  
 ### style
 
 ```tsx
 SymbolLayerStyleProps
 ```
-FIX ME NO DESCRIPTION
+_required_
+Customizable style attributes
