+++ src/components/CircleLayer.tsx
@@ -7,9 +7,12 @@ import RNMBXCircleLayerNativeComponent from '../specs/RNMBXCircleLayerNativeComp
 
 import AbstractLayer from './AbstractLayer';
 
-const MapboxGL = NativeModules.RNMBXModule;
+const Mapbox = NativeModules.RNMBXModule;
 
-export type Props = {
+// @{codepart-replace-start(LayerPropsCommon.codepart-tsx)}
+type Slot = 'bottom' | 'middle' | 'top';
+
+type LayerPropsCommon = {
   /**
    * A string that uniquely identifies the source in the style to which it is added.
    */
@@ -28,8 +31,7 @@ export type Props = {
   sourceID?: string;
 
   /**
-   * Identifier of the layer within the source identified by the sourceID property
-   * from which the receiver obtains the data to style.
+   * Identifier of the layer within the source identified by the sourceID property from which the receiver obtains the data to style.
    */
   sourceLayerID?: string;
 
@@ -64,6 +66,16 @@ export type Props = {
   maxZoomLevel?: number;
 
   /**
+   * The slot this layer is assigned to. If specified, and a slot with that name exists, it will be placed at that position in the layer order.
+   *
+   * v11 only
+   */
+  slot?: Slot;
+};
+// @{codepart-replace-end}
+
+export type Props = LayerPropsCommon & {
+  /**
    * Customizable style attributes
    */
   style?: CircleLayerStyleProps;
@@ -78,7 +90,7 @@ type NativeTypeProps = Omit<Props, 'style'> & {
  */
 class CircleLayer extends AbstractLayer<Props, NativeTypeProps> {
   static defaultProps = {
-    sourceID: MapboxGL.StyleSource.DefaultSourceID,
+    sourceID: Mapbox.StyleSource.DefaultSourceID,
   };
 
   render() {
